+++ src/fscript_ext/fscript_endian.c
@@ -80,20 +80,22 @@ static ret_t func_ntohll(fscript_t* fscript, fscript_args_t* args, value_t* resu
   return RET_OK;
 }
 
+FACTORY_TABLE_BEGIN(s_ext_endian)
+  FACTORY_TABLE_ENTRY("is_little", func_is_little_endian)
+
+  FACTORY_TABLE_ENTRY("htonl", func_htonl)
+  FACTORY_TABLE_ENTRY("ntohl", func_ntohl)
+
+  FACTORY_TABLE_ENTRY("htons", func_htons)
+  FACTORY_TABLE_ENTRY("ntohs", func_ntohs)
+
+  FACTORY_TABLE_ENTRY("htonf", func_htonf)
+  FACTORY_TABLE_ENTRY("ntohf", func_ntohf)
+
+  FACTORY_TABLE_ENTRY("htonll", func_htonll)
+  FACTORY_TABLE_ENTRY("ntohll", func_ntohll)
+FACTORY_TABLE_END()
+
 ret_t fscript_endian_register(void) {
-  ENSURE(fscript_register_func("is_little", func_is_little_endian) == RET_OK);
-
-  ENSURE(fscript_register_func("htonl", func_htonl) == RET_OK);
-  ENSURE(fscript_register_func("ntohl", func_ntohl) == RET_OK);
-
-  ENSURE(fscript_register_func("htons", func_htons) == RET_OK);
-  ENSURE(fscript_register_func("ntohs", func_ntohs) == RET_OK);
-
-  ENSURE(fscript_register_func("htonf", func_htonf) == RET_OK);
-  ENSURE(fscript_register_func("ntohf", func_ntohf) == RET_OK);
-
-  ENSURE(fscript_register_func("htonll", func_htonll) == RET_OK);
-  ENSURE(fscript_register_func("ntohll", func_ntohll) == RET_OK);
-
-  return RET_OK;
+  return fscript_register_funcs(s_ext_endian);
 }