+++ src/fscript_ext/fscript_iostream.c
@@ -39,10 +39,12 @@ static ret_t func_iostream_get_ostream(fscript_t* fscript, fscript_args_t* args,
 
   return RET_OK;
 }
+  
+FACTORY_TABLE_BEGIN(s_ext_iostream)
+  FACTORY_TABLE_ENTRY("iostream_get_istream", func_iostream_get_istream)
+  FACTORY_TABLE_ENTRY("iostream_get_ostream", func_iostream_get_ostream)
+FACTORY_TABLE_END()
 
 ret_t fscript_iostream_register(void) {
-  ENSURE(fscript_register_func("iostream_get_istream", func_iostream_get_istream) == RET_OK);
-  ENSURE(fscript_register_func("iostream_get_ostream", func_iostream_get_ostream) == RET_OK);
-
-  return RET_OK;
+  return fscript_register_funcs(s_ext_iostream);
 }