+++ src/fscript_ext/fscript_iostream_file.c
@@ -49,8 +49,11 @@ static ret_t func_ostream_file_create(fscript_t* fscript, fscript_args_t* args, 
   return RET_OK;
 }
 
+FACTORY_TABLE_BEGIN(s_ext_iostream_file)
+  FACTORY_TABLE_ENTRY("istream_file_create", func_istream_file_create)
+  FACTORY_TABLE_ENTRY("ostream_file_create", func_ostream_file_create)
+FACTORY_TABLE_END()
+
 ret_t fscript_iostream_file_register(void) {
-  ENSURE(fscript_register_func("istream_file_create", func_istream_file_create) == RET_OK);
-  ENSURE(fscript_register_func("ostream_file_create", func_ostream_file_create) == RET_OK);
-  return RET_OK;
+  return fscript_register_funcs(s_ext_iostream_file);
 }