+++ src/fscript_ext/fscript_iostream_inet.c
@@ -53,8 +53,11 @@ static ret_t func_iostream_udp_create(fscript_t* fscript, fscript_args_t* args, 
   return RET_OK;
 }
 
+FACTORY_TABLE_BEGIN(s_ext_iostream_inet)
+  FACTORY_TABLE_ENTRY("iostream_tcp_create", func_iostream_tcp_create)
+  FACTORY_TABLE_ENTRY("iostream_udp_create", func_iostream_udp_create)
+FACTORY_TABLE_END()
+
 ret_t fscript_iostream_inet_register(void) {
-  ENSURE(fscript_register_func("iostream_tcp_create", func_iostream_tcp_create) == RET_OK);
-  ENSURE(fscript_register_func("iostream_udp_create", func_iostream_udp_create) == RET_OK);
-  return RET_OK;
+  return fscript_register_funcs(s_ext_iostream_inet);
 }