+++ src/fscript_ext/fscript_json.c
@@ -101,11 +101,13 @@ static ret_t func_ubjson_save(fscript_t* fscript, fscript_args_t* args, value_t*
   return func_json_save_ex(fscript, args, result, conf_ubjson_save_as);
 }
 
-ret_t fscript_json_register(void) {
-  ENSURE(fscript_register_func("json_load", func_json_load) == RET_OK);
-  ENSURE(fscript_register_func("json_save", func_json_save) == RET_OK);
-  ENSURE(fscript_register_func("ubjson_load", func_ubjson_load) == RET_OK);
-  ENSURE(fscript_register_func("ubjson_save", func_ubjson_save) == RET_OK);
+FACTORY_TABLE_BEGIN(s_ext_json)
+  FACTORY_TABLE_ENTRY("json_load", func_json_load)
+  FACTORY_TABLE_ENTRY("json_save", func_json_save)
+  FACTORY_TABLE_ENTRY("ubjson_load", func_ubjson_load)
+  FACTORY_TABLE_ENTRY("ubjson_save", func_ubjson_save)
+FACTORY_TABLE_END()
 
-  return RET_OK;
+ret_t fscript_json_register(void) {
+  return fscript_register_funcs(s_ext_json);
 }