+++ src/fscript_ext/fscript_typed_array.c
@@ -187,16 +187,18 @@ static ret_t func_typed_array_join(fscript_t* fscript, fscript_args_t* args, val
   return RET_OK;
 }
 
-ret_t fscript_typed_array_register(void) {
-  ENSURE(fscript_register_func("typed_array_create", func_typed_array_create) == RET_OK);
-  ENSURE(fscript_register_func("typed_array_push", func_typed_array_push) == RET_OK);
-  ENSURE(fscript_register_func("typed_array_pop", func_typed_array_pop) == RET_OK);
-  ENSURE(fscript_register_func("typed_array_get", func_typed_array_get) == RET_OK);
-  ENSURE(fscript_register_func("typed_array_set", func_typed_array_set) == RET_OK);
-  ENSURE(fscript_register_func("typed_array_insert", func_typed_array_insert) == RET_OK);
-  ENSURE(fscript_register_func("typed_array_remove", func_typed_array_remove) == RET_OK);
-  ENSURE(fscript_register_func("typed_array_clear", func_typed_array_clear) == RET_OK);
-  ENSURE(fscript_register_func("typed_array_join", func_typed_array_join) == RET_OK);
+FACTORY_TABLE_BEGIN(s_ext_typed_array)
+  FACTORY_TABLE_ENTRY("typed_array_create", func_typed_array_create)
+  FACTORY_TABLE_ENTRY("typed_array_push", func_typed_array_push)
+  FACTORY_TABLE_ENTRY("typed_array_pop", func_typed_array_pop)
+  FACTORY_TABLE_ENTRY("typed_array_get", func_typed_array_get)
+  FACTORY_TABLE_ENTRY("typed_array_set", func_typed_array_set)
+  FACTORY_TABLE_ENTRY("typed_array_insert", func_typed_array_insert)
+  FACTORY_TABLE_ENTRY("typed_array_remove", func_typed_array_remove)
+  FACTORY_TABLE_ENTRY("typed_array_clear", func_typed_array_clear)
+  FACTORY_TABLE_ENTRY("typed_array_join", func_typed_array_join)
+FACTORY_TABLE_END()
 
-  return RET_OK;
+ret_t fscript_typed_array_register(void) {
+  return fscript_register_funcs(s_ext_typed_array);
 }