+++ tests/fscript_run.cpp
@@ -58,8 +58,10 @@ static ret_t run_fscript_file(const char* filename, uint32_t times) {
 int main(int argc, char* argv[]) {
   platform_prepare();
 
+  tk_mem_dump();
   fscript_global_init();
   fscript_ext_init();
+  tk_mem_dump();
   data_writer_factory_set(data_writer_factory_create());
   data_reader_factory_set(data_reader_factory_create());
   data_writer_factory_register(data_writer_factory(), "file", data_writer_file_create);
@@ -67,6 +69,7 @@ int main(int argc, char* argv[]) {
   data_reader_factory_register(data_reader_factory(), "asset", data_reader_asset_create);
   data_reader_factory_register(data_reader_factory(), "mem", data_reader_mem_create);
   data_writer_factory_register(data_writer_factory(), "wbuffer", data_writer_wbuffer_create);
+  tk_mem_dump();
 
   app_conf_init_json("runFScript");
   tk_mem_dump();